Completely redo the structure of class-related objects and database management
From now on, the structure can be represented like so:

Class -> ClassDetails -> ClassTimes

That is to say, every Class consists of an associated subject id and ClassDetail ids. Perhaps soon, each class can have a unit as well (e.g. Unit 3).

A ClassDetail has data about the room of the lesson and the teacher teaching it. Since each class may have multiple rooms and teachers, a Class contains a list of the ClassDetail ids. Also, each ClassDetail has multiple ClassTime objects associated (via a list of ids again).

A ClassTime object merely stores information about the day of the lesson and the start and end times. One may have multiple lessons of the same class with the same teacher and in the same room which is why multiple instances of this object can be associated with a ClassDetail.

While, this new structure results in more work to fetch particular values, I believe that its advantage outweighs this disadvantage; this structure allows for greater flexibility when adding or editing a new class. This will also be reflected in the layout for adding/editing these which I will add in the near future at the time of writing this commit description.