Heavily refactor to avoid passing DataHandler - use inheritance instead
`DataUtils` is now an interface instead of an object. This allows us to implement it in other classes (we created new classes called AssignmentUtils, ClassDetailUtils, etc.), so that we can inherit the default behaviours for actions such as getAllItems(). These new util classes each have to specify fields like tableName, idCol, etc., and this means we can avoid having to pass DataHandler, since the util class already has this data for us.

There are currently a lot of errors, so these need to be fixed in the next few commits!