Add a welcome page for configuring week scheduling
Note that we've also changed the setting for displaying weeks as letters to displaying weeks as numbers (i.e. it's been reversed). This is because it makes more sense to stick with Week A, Week B, etc. as the default instead of Week 1, Week 2, etc. Therefore, the functions in PrefUtils have been renamed and the string preference values and preference keys have also been changed accordingly.