Rewrite MainActivity in Kotlin and dynamically generate layout
The layout is much better designed now, where MainActivity shows two tabs: 'Today' and 'Upcoming'. Only the code for the 'Today' tab has been added up to this point.

Note the SectionUi class - which contains a Builder used to help generate sections of the UI dynamically.

In this commit, we've also deleted a lot of useless classes: adapter classes and all 'card' classes for the previous home screen.

We've also implemented Comparable in a number of framework model classes, defining the natural sorting order of these objects.